+++ database/rrdhost.c
@@ -64,8 +64,8 @@ static inline void rrdhost_init() {
 // ----------------------------------------------------------------------------
 // RRDHOST index by UUID
 
-inline long rrdhost_hosts_available(void) {
-    return (long)dictionary_entries(rrdhost_root_index);
+inline size_t rrdhost_hosts_available(void) {
+    return dictionary_entries(rrdhost_root_index);
 }
 
 inline RRDHOST *rrdhost_find_by_guid(const char *guid) {
